Issue #305 - Rename CanCatch, CanHandleError and CanRecover to canCatch, canHandleError and canRecover respectively
