+++ cats-effect/src/main/scala-3/effectie/cats/canCatch.scala
@@ -4,21 +4,20 @@ import cats.Id
 import cats.data.EitherT
 import cats.effect.IO
 import cats.syntax.all.*
+import effectie.core.CanCatch
 
 import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-06-07
   */
-object CanCatch {
-
-  type CanCatch[F[*]] = effectie.core.CanCatch[F]
+object canCatch {
 
   given canCatchIo: CanCatch[IO] with {
 
     inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
-    override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
+    inline override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
       fa.attempt
 
   }
@@ -27,7 +26,7 @@ object CanCatch {
 
     inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
 
-    override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
+    inline override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
       scala.util.Try(fa) match {
         case scala.util.Success(a) =>
           a.asRight[Throwable]
