+++ cats-effect3/src/main/scala-2/effectie/cats/canCatch.scala
@@ -3,18 +3,18 @@ package effectie.cats
 import cats.Id
 import cats.effect.IO
 import cats.syntax.all._
+import effectie.core.CanCatch
 
 /** @author Kevin Lee
   * @since 2020-06-07
   */
-object CanCatch {
-  private type CanCatch[F[_]] = effectie.core.CanCatch[F]
+object canCatch {
 
   implicit object CanCatchIo extends CanCatch[IO] {
 
     @inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
-    override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
+    @inline override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
       fa.attempt
 
   }
@@ -23,7 +23,7 @@ object CanCatch {
 
     @inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
 
-    override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
+    @inline override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
       scala.util.Try(fa) match {
         case scala.util.Success(a) =>
           a.asRight[Throwable]
