+++ cats-effect3/src/main/scala-3/effectie/cats/fx.scala
@@ -21,19 +21,19 @@ object fx {
     inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
     inline override final def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
-      CanCatch.canCatchIo.catchNonFatalThrowable(fa)
+      canCatch.canCatchIo.catchNonFatalThrowable(fa)
 
     inline override final def handleNonFatalWith[A, AA >: A](fa: => IO[A])(handleError: Throwable => IO[AA]): IO[AA] =
-      CanHandleError.ioCanHandleError.handleNonFatalWith(fa)(handleError)
+      canHandleError.ioCanHandleError.handleNonFatalWith(fa)(handleError)
 
     inline override final def handleNonFatal[A, AA >: A](fa: => IO[A])(handleError: Throwable => AA): IO[AA] =
-      CanHandleError.ioCanHandleError.handleNonFatal(fa)(handleError)
+      canHandleError.ioCanHandleError.handleNonFatal(fa)(handleError)
 
     inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => IO[A])(handleError: PartialFunction[Throwable, IO[AA]]): IO[AA] =
-      CanRecover.ioCanRecover.recoverFromNonFatalWith(fa)(handleError)
+      canRecover.ioCanRecover.recoverFromNonFatalWith(fa)(handleError)
 
     inline override final def recoverFromNonFatal[A, AA >: A](fa: => IO[A])(handleError: PartialFunction[Throwable, AA]): IO[AA] =
-      CanRecover.ioCanRecover.recoverFromNonFatal(fa)(handleError)
+      canRecover.ioCanRecover.recoverFromNonFatal(fa)(handleError)
 
   }
 
@@ -50,19 +50,19 @@ object fx {
     inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
 
     inline override final def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
-      CanCatch.canCatchId.catchNonFatalThrowable(fa)
+      canCatch.canCatchId.catchNonFatalThrowable(fa)
 
     inline override final def handleNonFatalWith[A, AA >: A](fa: => Id[A])(handleError: Throwable => Id[AA]): Id[AA] =
-      CanHandleError.idCanHandleError.handleNonFatalWith(fa)(handleError)
+      canHandleError.idCanHandleError.handleNonFatalWith(fa)(handleError)
 
     inline override final def handleNonFatal[A, AA >: A](fa: => Id[A])(handleError: Throwable => AA): Id[AA] =
-      CanHandleError.idCanHandleError.handleNonFatal(fa)(handleError)
+      canHandleError.idCanHandleError.handleNonFatal(fa)(handleError)
 
     inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => Id[A])(handleError: PartialFunction[Throwable, Id[AA]]): Id[AA] =
-      CanRecover.idCanRecover.recoverFromNonFatalWith(fa)(handleError)
+      canRecover.idCanRecover.recoverFromNonFatalWith(fa)(handleError)
 
     inline override final def recoverFromNonFatal[A, AA >: A](fa: => Id[A])(handleError: PartialFunction[Throwable, AA]): Id[AA] =
-      CanRecover.idCanRecover.recoverFromNonFatal(fa)(handleError)
+      canRecover.idCanRecover.recoverFromNonFatal(fa)(handleError)
 
   }
