+++ cats-effect3/src/test/scala-3/effectie/cats/syntax/errorSpec.scala
@@ -9,7 +9,7 @@ import effectie.cats.CatsEffectRunner.TestContext
 import effectie.cats.compat.CatsEffectIoCompatForFuture
 import effectie.cats.syntax.error.*
 import effectie.core.FxCtor
-import effectie.cats.{CanCatch, testing}
+import effectie.cats.{canCatch, testing}
 import effectie.syntax.error.*
 import effectie.syntax.fx.*
 import effectie.testing.types.*
@@ -1112,7 +1112,7 @@ object CanHandleErrorSyntaxSpec {
     effectOf[F](a)
 
   object IoSpec {
-    import effectie.cats.CanHandleError.ioCanHandleError
+    import effectie.cats.canHandleError.ioCanHandleError
     import effectie.cats.fx.given
 
     def testCanHandleError_IO_handleNonFatalWithShouldHandleNonFatalWith: Result = {
@@ -2027,7 +2027,7 @@ object CanHandleErrorSyntaxSpec {
   }
 
   object IdSpec {
-    import effectie.cats.CanHandleError.idCanHandleError
+    import effectie.cats.canHandleError.idCanHandleError
     import effectie.cats.fx.given
 
     def testCanHandleError_Id_handleNonFatalWithShouldHandleNonFatalWith: Result = {
@@ -2772,7 +2772,7 @@ object CanRecoverSyntaxSpec {
     effectOf[F](a)
 
   object IOSpec {
-    import effectie.cats.CanRecover.given
+    import effectie.cats.canRecover.given
     import effectie.cats.fx.given
 
     def testCanRecover_IO_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
@@ -3783,7 +3783,7 @@ object CanRecoverSyntaxSpec {
   }
 
   object IdSpec {
-    import effectie.cats.CanRecover.given
+    import effectie.cats.canRecover.given
     import effectie.cats.fx.given
 
     def testCanRecover_Id_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
