+++ effectie-monix/src/main/scala/effectie/monix/canCatch.scala
@@ -3,19 +3,19 @@ package effectie.monix
 import cats.Id
 import cats.effect.IO
 import cats.syntax.all._
+import effectie.core.CanCatch
 import monix.eval.Task
 
 /** @author Kevin Lee
   * @since 2020-06-07
   */
-object CanCatch {
-  private type CanCatch[F[_]] = effectie.core.CanCatch[F]
+object canCatch {
 
   implicit object CanCatchTask extends CanCatch[Task] {
 
     @inline override final def mapFa[A, B](fa: Task[A])(f: A => B): Task[B] = fa.map(f)
 
-    override def catchNonFatalThrowable[A](fa: => Task[A]): Task[Either[Throwable, A]] =
+    @inline override def catchNonFatalThrowable[A](fa: => Task[A]): Task[Either[Throwable, A]] =
       fa.attempt
 
   }
@@ -24,7 +24,7 @@ object CanCatch {
 
     @inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
-    override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
+    @inline override def catchNonFatalThrowable[A](fa: => IO[A]): IO[Either[Throwable, A]] =
       fa.attempt
 
   }
@@ -33,7 +33,7 @@ object CanCatch {
 
     @inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
 
-    override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
+    @inline override def catchNonFatalThrowable[A](fa: => Id[A]): Id[Either[Throwable, A]] =
       scala.util.Try(fa) match {
         case scala.util.Success(a) =>
           a.asRight[Throwable]
