+++ effectie-monix/src/main/scala/effectie/monix/canHandleError.scala
@@ -2,6 +2,7 @@ package effectie.monix
 
 import cats.Id
 import cats.effect.IO
+import effectie.core.CanHandleError
 import monix.eval.Task
 
 import scala.util.control.NonFatal
@@ -9,9 +10,7 @@ import scala.util.control.NonFatal
 /** @author Kevin Lee
   * @since 2020-08-17
   */
-object CanHandleError {
-
-  type CanHandleError[F[_]] = effectie.core.CanHandleError[F]
+object canHandleError {
 
   implicit object TaskCanHandleError extends CanHandleError[Task] {
